moved temporary storage switch into get started tab directory
